requirements.txt
web3==6.11.1
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.4.2
scikit-learn==1.3.2
pandas==2.1.2
numpy==1.26.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
aiohttp==3.8.6
tensorflow==2.14.0